requirements.txt
Django==2.1.1
Pillow==5.2.0
django-star-ratings==0.6.1
django-filter==2.0.0
django-mptt==0.9.1
celery==4.2.1
django-allauth==0.37.1
djangorestframework==3.8.2
Markdown==3.0.1
django-rest-swagger==2.2.0